Add ability to read ad engine version from ads context
