--- app/modules/is-ad-engine-experimental.js
@@ -1,12 +1,0 @@
-function isAdEngineExperimental() {
-  try {
-    const urlParams = new URLSearchParams(window.location.search);
-    const param = urlParams.get('adengine_experimental');
-
-    return param === '1';
-  } catch (e) {
-    return false;
-  }
-}
-
-export default isAdEngineExperimental;
